% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_qu_path_script.R
\name{generate_qupath_script}
\alias{generate_qupath_script}
\title{Generate the groovy script used for the analysis}
\usage{
generate_qupath_script()
}
\value{
\verb{script_for_qupath.txt in local dir}.
}
\description{
Generate a useful script to consistently save the output data from QuPath in .csv format following the naming conventions
followed during the package development.
}
