% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_binding.R
\name{process_file}
\alias{process_file}
\title{Process a Single File}
\usage{
process_file(file_path, extension)
}
\arguments{
\item{file_path}{Path to the file}

\item{extension}{String File extension to filter}
}
\value{
dataframe
}
\description{
This function returns a processed single file
}
\keyword{internal}
