% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{string_parsing}
\alias{string_parsing}
\title{Main parsing function}
\usage{
string_parsing(.data)
}
\arguments{
\item{.data}{dataframe with parsed metadata}
}
\value{
A \code{dataframe}/\code{tibble}.
}
\description{
This function will parse the data from the Image name and will return the metadata there contained
The metadata will be then associated to the count file as well
}
