% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_binding.R
\name{utils_internal}
\alias{utils_internal}
\alias{list_all_files}
\title{Internal utility functions for file handling}
\usage{
list_all_files(define_path, extension, recursive_search)
}
\description{
This file contains internal utility functions for file handling and processing
}
\keyword{internal}
