% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_binding.R
\name{data_binding}
\alias{data_binding}
\title{Merge all the dataframes coming out from the QuPath}
\usage{
data_binding(
  path_to_the_projects_folder,
  files_extension_to_look_for = "csv",
  recursive_search = FALSE
)
}
\arguments{
\item{path_to_the_projects_folder}{String/Path The path where the files coming out of QuPath are located}

\item{files_extension_to_look_for}{String The extension of the file outputted from QuPath, (default is "csv")}

\item{recursive_search}{Boolean, it defined the behavior of the file search, if recursive or not, (default is FALSE)}
}
\value{
A \code{dataframe}/\code{tibble}.
}
\description{
This function try to guess the string patterns that are in the dataset and then fill the dataframe
with that information. Finally the data is combined and combined them into one file
}
