% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_QC_plots_and_stats.R
\name{get_QC_plots_parsed_merged_data}
\alias{get_QC_plots_parsed_merged_data}
\title{Plot some QC plots for the bound data}
\usage{
get_QC_plots_parsed_merged_data(
  .data,
  list_of_columns_to_plot = NULL,
  save_plots = FALSE,
  saving_plots_folder = "figures",
  save_plots_in_patient_specific_subfolders = TRUE,
  fill_color_variable = NULL,
  PID_column_name = "PID",
  isolate_specific_drug = NULL,
  isolate_specific_patient = NULL,
  drug_column_name = "Treatment",
  save_list_of_plots = TRUE,
  p_height = 10,
  p_width = 10
)
}
\arguments{
\item{.data}{The preprocessed data (after running make_count_dataframe() and change_data_format_to_longer()) merged data.frame that should be visualized}

\item{list_of_columns_to_plot}{The preprocessed data (after running make_count_dataframe() and change_data_format_to_longer()) merged data.frame that should be visualized}

\item{save_plots}{Boolean, TRUE if plots should be saved (default is FALSE)}

\item{saving_plots_folder}{String indicating the folder where the plots should be stored (default is "figures")}

\item{save_plots_in_patient_specific_subfolders}{Boolean, if TRUE the plots will be saved (if \code{save_plots} TRUE) in sample specific folders (default is TRUE)}

\item{fill_color_variable}{Boolean, String, indicating the name of the variable (discrete) to use for the plot's filling}

\item{PID_column_name}{String, indicating the name of the sample to subset (default is "Treatment")}

\item{isolate_specific_drug}{String, indicating if there should be a Treatment specific data subset (default is NULL)}

\item{isolate_specific_patient}{String, indicating a spacific sample to plot only (default is NULL)}

\item{drug_column_name}{String, indicate the column indicating the Drug/Treament (default is "Treatment")}

\item{save_list_of_plots}{Boolean, if TRUE returns a named list of all the plots ran (default is TRUE), this can be usefult to isolate specific plots}

\item{p_height}{Integer, indicate the plot's height (default is 10 inches)}

\item{p_width}{Integer, indicate the plot's width (default is 10 inches)}
}
\value{
A \code{list}/\code{NULL}.
}
\description{
This plot can show trends within the dataset and run some basic statistics.
}
