% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_run_config.R
\name{make_run_config}
\alias{make_run_config}
\title{Generates and use a config txt file}
\usage{
make_run_config(overwrite_config = FALSE, forcePath = NULL)
}
\arguments{
\item{overwrite_config}{Boolean, if TRUE the \code{config_drugsens.txt} will be overwritten (default is FALSE)}

\item{forcePath}{String, Define a custom path for the config file}
}
\value{
A \code{dataframe}/\code{tibble}.
}
\description{
When this function run the first time, it will generated a config.txt file in the user working directory.
It will import the data config file into the use environment. This data will be used to change the column names
of the imported dataset and change the name of the markers that is often incorrectly exported.
}
\examples{
\dontrun{
  make_run_config()
}
}
